+++ docs/CHANGELOG.md
@@ -6,6 +6,22 @@
 
 * feat: adds support for verified queries against management canister
   * includes support for `fetch_canister_logs` in the actor provided by `getManagementCanister`
+  * allows you to pass an `effectiveCanisterId` to the `getManagementCanister` api, which makes it possible to query bitcoin methods as well. For example:
+
+```ts
+// For now, the verifyQuerySignatures option must be set to false
+const agent = await makeAgent({ host: 'https://icp-api.io', verifyQuerySignatures: false });
+const management = getManagementCanister({
+    agent,
+    effectiveCanisterId: Principal.from('mm444-5iaaa-aaaar-qaabq-cai'),
+});
+
+const result = await management.bitcoin_get_balance_query({
+    address: 'bc1qxy2kgdygjrsqtzq2n0yrf2493p83kkfjhx0wlh',
+    network: { mainnet: null },
+    min_confirmations: [6],
+});
+```
 
 ### Changed